% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rollingCentralMoment}
\alias{rollingCentralMoment}
\alias{rollingMax}
\alias{rollingMean}
\alias{rollingMedian}
\alias{rollingMin}
\alias{rollingNobs}
\alias{rollingProduct}
\alias{rollingSD}
\alias{rollingSum}
\alias{rollingSumStable}
\alias{rollingVar}
\title{Rolling operations functions for irregularly spaced time series}
\usage{
rollingCentralMoment(times, values, widthbefore, widthafter, moment)

rollingMax(times, values, widthbefore, widthafter)

rollingMean(times, values, widthbefore, widthafter)

rollingMedian(times, values, widthbefore, widthafter)

rollingMin(times, values, widthbefore, widthafter)

rollingNobs(times, values, widthbefore, widthafter)

rollingProduct(times, values, widthbefore, widthafter)

rollingSD(times, values, widthbefore, widthafter)

rollingSum(times, values, widthbefore, widthafter)

rollingSumStable(times, values, widthbefore, widthafter)

rollingVar(times, values, widthbefore, widthafter)
}
\arguments{
\item{times}{A Datetime vector}

\item{values}{A numeric vector}

\item{widthbefore}{A double with the preceding observation width}

\item{widthafter}{A double with the subsequent observation width}

\item{moment}{A double with the requested moment.}
}
\value{
A numeric vector with the corresponding result.
}
\description{
The UTS library by Andreas Eckner provides algorithms for unevenly
spaced time-series data.  This package brings a few of them to R.
The functions describe here offer various rolling operators.
}
\author{
Dirk Eddelbuettel for the package, Andreas Eckner for the
underlying code.
}
