% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SMAnext}
\alias{SMAnext}
\alias{SMAlast}
\alias{SMAlinear}
\title{SMA functions for irregularly samples time series}
\usage{
SMAnext(times, values, widthbefore, widthafter)

SMAlast(times, values, widthbefore, widthafter)

SMAlinear(times, values, widthbefore, widthafter)
}
\arguments{
\item{times}{A Datetime vector}

\item{values}{A numeric vector}

\item{widthbefore}{A double with the preceding observation width}

\item{widthafter}{gvA double with the subsequent observation width}
}
\value{
A numeric vector with SMA-weighted values.
package at the given position is available.
}
\description{
The UTS library by Andreas Eckner provides algorithms for unevenly
spaced time-series data.  This package brings a few of them to R;
and these functions offer exponentially-decaying weighted moving
average, or SMA, for short. Three variants are provides considering
the last or next observation relative to time \sQuote{t}, as well as
linear interpolation between them.
}
\examples{
tt <- Sys.time() + c(0, 0.2, 1.2, 2.3)
vv <- 0:3
SMAnext(tt, vv, 2.5, 1.5)
SMAlast(tt, vv, 2.5, 1.5)
SMAlinear(tt, vv, 2.5, 1.5)
}
\author{
Dirk Eddelbuettel for the package, Andreas Eckner for the
underlying code.
}
