% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{utsExample}
\alias{utsExample}
\title{Irregularly spaced time series example}
\usage{
utsExample()
}
\value{
Nothing
}
\description{
The UTS library by Andreas Eckner provides algorithms for unevenly
spaced time-series data.  This package brings a few of them to R.
This function shows the original example.
}
\examples{
utsExample()
}
\author{
Dirk Eddelbuettel for the package, Andreas Eckner for the
underlying code.
}
