% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{EMAnext}
\alias{EMAnext}
\alias{EMAlast}
\alias{EMAlinear}
\title{EMA functions for irregularly samples time series}
\usage{
EMAnext(times, values, tau)

EMAlast(times, values, tau)

EMAlinear(times, values, tau)
}
\arguments{
\item{times}{A Datetime vector}

\item{values}{A numeric vector}

\item{tau}{A double with the decay factor}
}
\value{
A numeric vector with EMA-weighted values.
package at the given position is available.
}
\description{
The UTS library by Andreas Eckner provides algorithms for unevenly
sampled time-series data.  This package brings a few of them to R;
and these functions offer exponentially-decaying weighted moving
average, or EMA, for short. Three variants are provides considering
the last or next observation relative to time \sQuote{t}, as well as
linear interpolation between them.
}
\examples{
tt <- Sys.time() + c(0, 0.2, 1.2, 2.3)
vv <- 0:3
EMAnext(tt, vv, 1)
EMAlast(tt, vv, 1)
EMAlinear(tt, vv, 1)
}
\author{
Dirk Eddelbuettel for the package, Andreas Eckner for the
underlying code.
}
