% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SMAnext}
\alias{SMAnext}
\alias{SMAlast}
\alias{SMAlinear}
\title{SMA functions for unevenly spaced time series}
\usage{
SMAnext(times, values, widthbefore, widthafter)

SMAlast(times, values, widthbefore, widthafter)

SMAlinear(times, values, widthbefore, widthafter)
}
\arguments{
\item{times}{A Datetime vector}

\item{values}{A numeric vector}

\item{widthbefore}{A double with the preceding observation width}

\item{widthafter}{gvA double with the subsequent observation width}
}
\value{
A numeric vector with SMA-weighted values.
package at the given position is available.
}
\description{
The UTS library by Andreas Eckner provides algorithms for unevenly
spaced time-series data.  This package brings a few of them to R.
The functions describe here offer simple moving
average, or SMA, for short. Three variants are provides considering
the last or next observation relative to time \sQuote{t}, as well as
linear interpolation between them.
}
\examples{
if (requireNamespace("xts", quietly=TRUE)) {
    suppressMessages(library(xts))
    times <- ISOdatetime(2018, 6, 7, 8, 30, 0) + c(0, 1.0, 1.2, 2.3, 2.9, 5.0)
    values <- seq(0, 10, by=2)
    plot(xts(values, order.by=times), type="b",
         main="Series and last/next/linear SMAs")
    lines(xts(SMAlast(times,values, 2.5, 1), values, order.by=times), 
         type="b", col="lightblue")
    lines(xts(SMAnext(times,values, 2.5, 1),  values, order.by=times),
         type="b", col="darkblue")
    lines(xts(SMAlinear(times,values, 2.5, 1),  values, order.by=times),
         type="b", col="mediumblue")
    addLegend("topleft", legend.names=c("series", "SMAlast", "SMAnext", "SMAlinear"),
              lty=rep(1,4), lwd=rep(1,4),
              col=c("black", "lightblue", "darkblue", "mediumblue"))
}
}
\author{
Dirk Eddelbuettel for the package, Andreas Eckner for the
underlying code.
}
